check clientId in activemq-plugin

+++ kalinka-activemq-plugin/src/main/java/org/diehl/dcs/kalinka/activemq_plugin/ActivemqPlugin.java
@@ -1,17 +1,10 @@
 /*
-Copyright [2017] [DCS <Info-dcs@diehl.com>]
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
+ * Copyright [2017] [DCS <Info-dcs@diehl.com>] Licensed under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance with the License. You may obtain a
+ * copy of the License at http://www.apache.org/licenses/LICENSE-2.0 Unless required by applicable
+ * law or agreed to in writing, software distributed under the License is distributed on an "AS IS"
+ * BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License
+ * for the specific language governing permissions and limitations under the License.
  */
 
 package org.diehl.dcs.kalinka.activemq_plugin;
@@ -35,12 +28,14 @@ public class ActivemqPlugin implements BrokerPlugin {
 
 	private static final Logger LOG = LoggerFactory.getLogger(ActivemqPlugin.class);
 
+	private static final String CLIENT_ID_PREFIX_KAFKA = "kafka-client";
+
 	private final ZkClient zkClient;
 	private final String host;
 
 	public ActivemqPlugin(final String zkServers, final String host) {
 
-		LOG.debug("Trying to create ZkClient for zkServers=" + zkServers + ", currentHost=" + host);
+		LOG.debug("Trying to create ZkClient for zkServers={}, currentHost={}", zkServers, host);
 		this.zkClient = new ZkClient(zkServers);
 		this.host = host;
 		LOG.info("Created ZkClient for zkServers=" + zkServers + ", currentHost=" + host);
@@ -56,15 +51,25 @@ public class ActivemqPlugin implements BrokerPlugin {
 			public void addConnection(final ConnectionContext context, final ConnectionInfo info)
 					throws Exception {
 
+				LOG.debug("Received connect from clientId={}", context.getClientId());
+
 				super.addConnection(context, info);
+				if (context.getClientId().startsWith(CLIENT_ID_PREFIX_KAFKA)) {
+					return;
+				}
 				upsertZkNode(context.getClientId());
 			}
 
 			@Override
 			public void removeConnection(final ConnectionContext context, final ConnectionInfo info,
-				final Throwable error) throws Exception {
+					final Throwable error) throws Exception {
+
+				LOG.debug("Received disconnect from clientId={}", context.getClientId());
 
 				super.removeConnection(context, info, error);
+				if (context.getClientId().startsWith(CLIENT_ID_PREFIX_KAFKA)) {
+					return;
+				}
 				deleteZkNode(context.getClientId());
 			}
 		};
@@ -76,8 +81,7 @@ public class ActivemqPlugin implements BrokerPlugin {
 		LOG.debug("Trying to upsert node={}, to host={}", node, this.host);
 		try {
 			this.zkClient.createPersistent(node, this.host);
-		}
-		catch(final ZkNodeExistsException e) {
+		} catch (final ZkNodeExistsException e) {
 			this.deleteZkNode(clientId);
 			this.zkClient.createPersistent(node, this.host);
 		}
@@ -91,5 +95,4 @@ public class ActivemqPlugin implements BrokerPlugin {
 		this.zkClient.delete(node);
 		LOG.info("Deleted node={}", node);
 	}
-
 }
